Created seperate menu components for LP and NVI (wip)
